update memo refactoring
