+++ guardian-service/src/hedera-modules/memo-mappings/message-memo.ts
@@ -85,7 +85,7 @@ export class MessageMemo extends MemoMap {
         messageMemo[`${MessageType.Topic}.${MessageAction.CreateTopic}.${TopicType.InstancePolicyTopic}`] = 'Policy Instance topic creation message';
         messageMemo[`${MessageType.Topic}.${MessageAction.CreateTopic}.${TopicType.UserTopic}`] = 'Standard Registry topic creation message';
         messageMemo[`${MessageType.Topic}.${MessageAction.CreateTopic}.${TopicType.DynamicTopic}`] = '${name} operation topic creation message';
-        messageMemo[`${MessageType.Token}.${MessageAction.UseToken}`] = 'UseToken';
+        messageMemo[`${MessageType.Token}.${MessageAction.UseToken}`] = 'Policy token issue message';
         messageMemo[MessageAction.ChangeMessageStatus] = 'Status change message';
         messageMemo[MessageAction.RevokeDocument] = 'Revoke document message';
         messageMemo[MessageAction.DeleteDocument] = 'Delete document message';
